fixed naming for token manager
