+++ src/cashier_frontend/src/components/manage-tokens/token-card.tsx
